refactor: SignInScreen
