+++ src/view/pages/SignInScreen/styles.ts
@@ -1,0 +1,24 @@
+import { StyleSheet } from 'react-native'
+
+export const styles = StyleSheet.create({
+  view: {
+    flex: 1,
+    paddingHorizontal: 10,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  divider: {
+    marginVertical: 5,
+  },
+  buttonsView: {
+    flexDirection: 'row',
+    justifyContent: 'space-evenly',
+  },
+  signInToSeaButton: {
+    flex: 1,
+  },
+  authButton: {
+    flex: 1,
+    marginLeft: 10,
+  },
+})
